+++ Frends.AzureBlobStorage.ListBlobsInContainer/Frends.AzureBlobStorage.ListBlobsInContainer.Tests/ConnectionStringUnitTests.cs
@@ -70,6 +70,8 @@ public class ConnectionStringUnitTests
             Assert.IsTrue(result.BlobList.Any(x => x.Name == "TestFile2.txt"));
             Assert.IsTrue(result.BlobList.Any(x => x.ETag != null));
             Assert.IsTrue(result.BlobList.Any(x => x.Type != null));
+            Assert.IsTrue(result.BlobList.Any(x => x.CreatedOn != null));
+            Assert.IsTrue(result.BlobList.Any(x => x.LastModified != null));
         }
     }
     
@@ -106,6 +108,8 @@ public class ConnectionStringUnitTests
             Assert.IsTrue(result.BlobList.Any(x => x.Name == "TestFile2.txt"));
             Assert.IsTrue(result.BlobList.Any(x => x.ETag != null));
             Assert.IsTrue(result.BlobList.Any(x => x.Type != null));
+            Assert.IsTrue(result.BlobList.Any(x => x.CreatedOn != null));
+            Assert.IsTrue(result.BlobList.Any(x => x.LastModified != null));
         }
     }