+++ Frends.AzureBlobStorage.ListBlobsInContainer/Frends.AzureBlobStorage.ListBlobsInContainer.Tests/OAuthUnitTests.cs
@@ -77,6 +77,8 @@ public class OAuthUnitTests
             Assert.IsTrue(result.BlobList.Any(x => x.Name == "TestFile2.txt"));
             Assert.IsTrue(result.BlobList.Any(x => x.ETag != null));
             Assert.IsTrue(result.BlobList.Any(x => x.Type != null));
+            Assert.IsTrue(result.BlobList.Any(x => x.CreatedOn != null));
+            Assert.IsTrue(result.BlobList.Any(x => x.LastModified != null));
         }
     }
     
@@ -116,6 +118,8 @@ public class OAuthUnitTests
             Assert.IsTrue(result.BlobList.Any(x => x.Name == "TestFile2.txt"));
             Assert.IsTrue(result.BlobList.Any(x => x.ETag != null));
             Assert.IsTrue(result.BlobList.Any(x => x.Type != null));
+            Assert.IsTrue(result.BlobList.Any(x => x.CreatedOn != null));
+            Assert.IsTrue(result.BlobList.Any(x => x.LastModified != null));
         }
     }