+++ Frends.AzureBlobStorage.ListBlobsInContainer/Frends.AzureBlobStorage.ListBlobsInContainer/Definitions/BlobData.cs
@@ -1,4 +1,6 @@
-﻿namespace Frends.AzureBlobStorage.ListBlobsInContainer.Definitions;
+﻿using System;
+
+namespace Frends.AzureBlobStorage.ListBlobsInContainer.Definitions;
 
 /// <summary>
 /// Blob's data.
@@ -28,4 +30,16 @@ public class BlobData
     /// </summary>
     /// <example>0x8D6FACDEB709651</example>
     public string ETag { get; set; }
+
+    /// <summary>
+    /// Blob's created on datetime.
+    /// </summary>
+    /// <example>2023-12-11T09:00:00.743185+02:00</example>
+    public DateTimeOffset? CreatedOn { get; set; }
+
+    /// <summary>
+    /// Blob's last modified datetime.
+    /// </summary>
+    /// <example>2023-12-11T09:00:00.743185+02:00</example>
+    public DateTimeOffset? LastModified { get; set; }
 }